examples/llm_prompt_optimization/wilds_active_learn_approach/requirements.txt
# Active Prompt Evolution - dependencies
# Install from project root: pip install -r requirements.txt

# Core (from wilds_experiment)
numpy
pyyaml
openai
python-dotenv
scikit-learn
sentence-transformers

# WILDS dataset
wilds

# OpenEvolve and its runtime deps (if not using pip install -e ../../..)
dacite>=1.9.2

# OpenEvolve (from project root)
# pip install -e ../../..  or  pip install -e openevolve/
